Extract cm6-mount + cm6-theme-sync into sections/*.js
First two region extractions of P-09. cm6-mount carries _cm state plus _mountCm / _destroyCm / _cmRetheme; cm6-theme-sync is a 9-LOC MutationObserver that imports _cmRetheme from cm6-mount. Keeping them as two files because a reviewer looking for "theme sync" wants a file named that.

New dashboard-main.js exports: _noteModal (object — ESM live-binding works because the importer mutates fields rather than reassigning), _setDirty, saveEdit. All three will follow their owning regions (note-preview-modal / note-mode) out in later P-09 commits.
